requirements.txt
streamlit
ccxt
psycopg2-binary
python-dotenv
bcrypt
pandas
aiohttp
